requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
Flask~=3.1.0
Flask-Cors~=5.0.0
pymongo~=4.10.1
redis~=5.2.1
bcrypt~=4.2.1
h11~=0.14.0
cryptography~=44.0.0
pip~=23.2.1
attrs~=24.3.0
PySocks~=1.7.1
cffi~=1.17.1
MarkupSafe~=3.0.2
typing_extensions~=4.12.2
trio~=0.28.0
outcome~=1.3.0.post0
sniffio~=1.3.1
sortedcontainers~=2.4.0
idna~=3.10
Werkzeug~=3.1.3
click~=8.1.8
Jinja2~=3.1.5
blinker~=1.9.0
itsdangerous~=2.2.0
certifi~=2024.12.14
tzlocal~=5.2
urllib3~=2.3.0
charset-normalizer~=3.4.1
selenium~=4.27.1
trio-websocket~=0.11.1
websocket-client~=1.8.0
packaging~=24.2
pycparser~=2.22
wsproto~=1.2.0
webdriver-manager~=4.0.2
PyJWT~=2.10.1
ics~=0.7.2
mailersend~=0.5.8